move model in its own file
